requirements.txt
-e git+git://github.com/django/django#egg=django
dj-database-url
django-mptt
django-tinymce==1.5.1a1
mimeparse==0.1.3
python-dateutil==1.5
python-digest==1.7
pil
django_compressor
django-celery
django-tastypie
celery-with-mongodb
django-social-auth
-e git+git://github.com/emperorcezar/django-recurrence#egg=recurrence
psycopg2
gunicorn
